gather: move helper functions to own utils.go; better fetcher tests
